Dev: Move 'types' folder into 'lib'
